// src/app/library/components/esri-widget/esri-widget-type.enum.ts
/**
 * Les types de widgets supportés par le socle js
 *
 * cf les sous-classes de __esri.Widget :
 *
 * @see https://developers.arcgis.com/javascript/latest/api-reference/esri-widgets-Widget.html
 */
export enum EsriWidgetType {
  AreaMeasurement3D = "esri.widget.AreaMeasurement3D",
  Attribution = "esri.widget.Attribution",
  BasemapGallery = "esri.widget.BasemapGallery",
  BasemapToggle = "esri.widget.BasemapToggle",
  Compass = "esri.widget.Compass",
  CoordinateConversion = "esri.widget.CoordinateConversion",
  DirectLineMeasurement3D = "esri.widget.DirectLineMeasurement3D",
  Directions = "esri.widget.Directions",
  Expand = "esri.widget.Expand",
  Feature = "esri.widget.Feature",
  Fullscreen = "esri.widget.Fullscreen",
  Home = "esri.widget.Home",
  LayerList = "esri.widget.LayerList",
  Legend = "esri.widget.Legend",
  Locate = "esri.widget.Locate",
  NavigationToggle = "esri.widget.NavigationToggle",
  Popup = "esri.widget.Popup",
  Print = "esri.widget.Print",
  ScaleBar = "esri.widget.ScaleBar",
  Search = "esri.widget.Search",
  Track = "esri.widget.Track",
  Zoom = "esri.widget.Zoom",
}